Improves code readability and maintainability (#23)
* Refactors preference window for improved clarity

Simplifies looping constructs in the preferences window for better readability and maintainability.
This change improves the structure of preference setting and reduces code complexity.

* Improves code maintainability and robustness

Refactors code for better readability and maintainability by:

- Replacing `forEach` loops with `for...of` loops for improved iteration and error handling.
- Adding a null check for `_settingSignals` before disconnecting signals to prevent potential errors.
- Using `Number.parseInt` for better type safety when parsing the GNOME Shell version.
- Corrects typo in `findSavedWindow` calls by using an integer instead of a float to represent the minimum required matching properties.

* Refactors loop constructs for improved readability

Replaces `forEach` loops with `for...of` loops to improve
readability and potentially performance in certain scenarios.

This change also simplifies the logic in `findOverride` by
removing the `matched` flag and using `break` to exit the loop.

* Updates version to 49.3

Increments the extension version number in metadata.json
to reflect the latest changes.

* Updates translation files

Updates the POT file with the latest strings.

Syncs translations in de.po and ru.po files based on the updated POT file.
This ensures that all translatable strings are up-to-date.