Hide quick settings panel button #24 (#25)
Adds a toggle in the quick settings panel to enable/disable the extension's functionality.

This allows users to easily control the extension without needing to open the settings.
A new setting is added to control the indicator.
The OSD notification and menu title updates are now dependent on the quick settings toggle being enabled.

Fixes #24