Enables configurable quick settings toggle (#43)
* Enables configurable quick settings toggle

Allows users to select the function of the quick settings toggle via a new preference setting.
This provides more flexibility and customization options for users to control
extension features directly from the quick settings menu.

Also updates the extension version to 49.11.

* Updates codeql action to v4

Updates the github/codeql-action/upload-sarif action to v4.

This ensures compatibility with the latest features and improvements in the CodeQL action, and uses latest security patches.

* Unbinds toggle before binding

The toggle is unbound before it is bound again.

This prevents errors and ensures that the toggle is correctly
bound to the setting.

* Refactors test assertions for clarity

Simplifies the test assertions in the common test file.
This improves readability and maintainability of the tests
by using `const` for immutable variables and removing
unnecessary line breaks.

* Simplifies score assertions in tests

Updates test assertions to use integer values instead of floats for better readability and consistency. This change improves the clarity of expected scores in various test cases.